requirements.txt
attrs==21.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.1.0
dnsq==1.1.6
expiringdict==1.2.1
filelock==3.4.0
flanker==0.9.11
future==0.18.2
grist-api==0.1.0
httplib2==0.20.2
idna==3.3
imap-tools==0.50.2
packaging==21.3
ply==3.11
pycparser==2.21
pyparsing==3.0.6
python-dotenv==0.19.2
redis==4.1.0
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
six==1.16.0
tld==0.12.6
tldextract==3.1.2
urllib3==1.26.7
WebOb==1.8.7
wrapt==1.13.3
